requirements_dev.txt
ipython==5.0.0
pytest-django==2.9.1
ipdb==0.10.1
pytest-mock==1.1
